Removing G\Introspection support for the generator
Pygobject is obviously in charge of the gobject introspection work
They introduced the concept of marking something as "introspectable=0"

This not only flags it as non-introspectable but removes it completely
from the resulting typelibs.  doesn't matter if other languages or tools
using gobject introspection typelib files might WANT those symbols
or find them completely introspectable

So it is impossible to use built typelibs or the extension for this
functionality.  Instead generated .gir xml files will be used